Add a post_install script that pushes out a copy of the system-level switcher settings to all the nodes after they have been booted for the first time.  This is necessary because the first time sync_users runs, it generates a checksum that is based on the switcher.ini that is generated after step 1.  The file never changes after this, so sync_users assumes that it does not need to be pushed to the nodes (even though it has never been pushed to the nodes in the first place).
I'm thinking that there used to be a "sync_users --force" somewhere in the
install process, to guarantee to sync users (and any other files that
sync_users is maintaining), but it doesn't seem to exist anymore.  So
this is a workaround to get the switcher settings out into the nodes
in the meantime.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@1096 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: b704ed0ecad34f62227434fcc12e4c38887390b0