First cut at a pre_configure script that generates an HTML page for the configurator step based on what packages are going to be installed, and if there are multiple packages attempting to use the same tag (e.g., LAM and MPICH).  When this happens, the user will be prompted with a choice.
Still need some kind of mechanism for a default selection in these cases.

Also, the input data to know what packages is currently hard-coded (although
it is available in the packages' config.xml files).  Waiting for ODA API
from Neil to ditch the hard-coding and retrieve it properly at run-time.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@1310 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: ed7ee001a0a6a48bbe783e4506366593114ba271