Add robust functionality to handle the combination of statements under locking and normal statements. For example, database_execute_command(...); database_execute_command(...); database_execute_command(...); ... locking(...); dec_already_locked(...); dec_already_locked(...); dec_already_locked(...); .. unlock(..); ..
database_execute_command(...);
database_execute_command(...);
database_execute_command(...);
..
After all, we are not going to use the database_execute_command but
just for now since we were not able to finish to apply locking to all
database_execute_commands, we still need to mix them.
If all dec's are locked, then this modification is not needed.
This modification is fine even in the case of only locking dec's.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@2620 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 39be5148e91827736722dcb99bf78fb2082005b0