Fixing the data inconsistency problem in deleting nodes. The detailed description that I emailed to oscar-core list is following. ----------------------------------------------------------------------------- Hi all,
I have been working on fixing the bug of DelNode.pm

There are two problems when a user clicked the "Delete OSCAR Clients.." button.

1) Even though it looks like the list of nodes is deleted in the "Delete OSCAR Clients" window, the oda database is not updated properly
2) delete_records, oda raw command, is not working with the record of NULL value.


The second problem is fixed by modifying @wheres in oda.pm, which tries to push any field names and values whether or not the value is NULL.

But the first problem is not fixed completely yet.
Thomas gave me the idea that the updated sis database data can only affect the sis/scripts/post_clients to update oda database.
So I think re-ordering commands in DelNode.pm can fix the inconsistency of database in the process of Delete OSCAR Clients.
As I expected, the oda database can be consistent by re-ordering the commands in DelNode.pm but the torque has an error in the process of post_install in the second trial of
deleting the second node.
For instance, deleting the first node is successfully done, but the second node is deleted but configuration is failed.
I have two client nodes. I have not tried to delete the two nodes at the same time yet.
I will try it and let you all know what happens.
--> I just did it and it works fine.
I think if a user deletes once, it works fine. But for the next one, it deletes nodes but the configuration is failed because of the above the problem.

The ordering of commands in DelNode.pm is...
1. Run post_clients without the one of sis pacakge
2. Run post_install
3. Turning off Client services.
4. Re-starting server services.
5. system("mksimachine --Delete --name $clientstring")
6. Run post_clients of sis packages.
7. Updating C3 configuration file\n";
8. Re-starting client services on remaining nodes

What I did is chang 1 and insert 6
-----------------------------------------------------------------------------


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-4-0@2734 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 2819bb5c73b3266134acb0dd387e727f8cebe5fa