Fix for bug 1149597:
Since the "man" command sucks (specifically: if the $MANPATH
environment variable is set, it does not look in /etc/man.config file
default man paths), and since the modules commands "append-path" and
"prepend-path" are too smart (meaning that they won't let us add an
empty path anywhere in MANPATH -- which is a pseudo-documented way of
making the "man" command read the default man paths from
/etc/man.config, even if $MANPATH is set), we have to add this
modulefile as a workaround.

The problem being solved is that if any of the OSCAR modulefiles
append-path or prepend-path to the MANPATH (which several of them do),
then all the default man pages cannot be found (e.g., "man ls").  A
previous workaround had been to directly set the MANPATH environment
variable to ":/opt/modules/default/man" (i.e., via setenv, not
prepend-path or append-path), which effectively did the "man"
workaround of an empty path in MANPATH, and therefore /etc/man.config
was consulted.  However, as soon as any modulefule uses prepend-path,
we lose the leading ":", and the empty path is gone.  End result: "man
ls" returns nothing.  Doh!

The only safe way to do this, therefore, is to read in all the paths
from /etc/man.config and add them to the $MANPATH -- we cannot
guarantee that the leading or trailing ":" will stay there forever.

This modulefile is non-switched, meaning that it will be loaded on
every shell, and is named default-manpath.  It will read in all the
MANPATH entries from /etc/man.config and add them to the MANPATH
environment variable.  If you don't want this to happen, then you can
just module unload default-manpath.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-4-1@2966 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: f8e18e2293ab5d7993d76133745583342614b021