$$database_nics_ref{ name } has not returned the reference of the hash table containing the records of nics table. I think this $$database_nics_ref{ name } is wrong originally to determine if there is no corresponding record in the database already since even though there are records for "name" in nics table, (i.e., after first installation), $$database_nics_ref{ name } shows that the return value does not exist, (i.e., there is no record for "name" in the nics table).
So, I changed the key for the $$database_nics_ref to determine the above status
correctly.
$$database_nics_ref{ name } ==> $$database_nics_ref{ $nic_name }
With respect to the test result of this on MDK 10, RHEL AS3, and FC2, 
it fixes the MySQL error during secondary installation on MDK 10.

But I would like to make sure if this modification breaks any code on the other
distros.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-4-1@3055 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: c81b5ba6637841c7e109b78c79db516e7e878d30