$$database_nics_ref{ name } has not returned the reference of the hash table containing the records of nics table. I think this $$database_nics_ref{ name } is wrong originally to determine if there is no corresponding record in the database since even though there are records for "name" in the nics table, (i.e., after first installation), $$database_nics_ref{ name } shows that the return value does not exist, (i.e., there is no record for "name" in the nics table).
So, I changed the key for the $$database_nics_ref to determine the above status
correctly.
$$database_nics_ref{ name } ==> $$database_nics_ref{ $nic_name }
With respect to the test result of this on MDK 10, RHEL AS3, and FC2, 
it fixes the MySQL error during secondary installation on MDK 10.

But I would like to make sure if this modification breaks any code on the other
distros.



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@3060 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: a5fac7580e9b2e27b96bc642571c0960bda8d807