Merge from trunk: svn merge -r3604:3605 https://svn.oscar.openclustergroup.org/svn/oscar/trunk svn merge -r3620:3621 https://svn.oscar.openclustergroup.org/svn/oscar/trunk svn merge -r3621:3622 https://svn.oscar.openclustergroup.org/svn/oscar/trunk
Fix for bug 1273950 : perl-Qt package selector error 
According to some communication between Bernard and
KDE developers this might result from a compiler issue on ia64.
The solution is to compile the code in smoke/qt with -O0 (only on ia64).
The code is not really performance relevant, so it's fine.
Code or compile options for other architectures isn't touched.

Rebuilt for other arches, too, for symmetry reasons.

FC2 and MDK packages still missing, but that's not critical, the old ones
should still work.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-4-2@3623 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: a9c7d2b495bb539b3d3e384e25abc6585df3bbd5