Yume is a small tool for integrating the wellknown package management tool yum (yellowdorg updater modified) into OSCAR. It simplifies the setup and management of heterogeneous clusters, i.e. clusters running multiple distributions or even multiple architectures controlled by one single master node.
This tool is very useful for managing clusters which are running distributions
using the rpm package format. It can:
 - prepare an rpm repository (i.e. generate a metadata cache in createrepo
   format, which is used also for apt, up2date and other package managers)
 - export it through apache, /tftpboot/rpm will be visible from cluster nodes
   as http://oscar_server/repo_rpm, including the metadata cache)
 - execute yum commands applying only to this repository (locally)
 - execute yum commands on the cluster nodes applying only to this repository.
This makes installing packages, creating cluster node images, updating
revisions much simpler than with rpm.

yume-opkg is a simple ODA query tool which facilitates installation and
deinstallation of opkgs into non-native images. To be able to fully exploit
it all packages will need to be converted to generic-setup format.

A tested yum package is included in this opkg. It is called yum-oscar and
should be used in distros which do not come with their own yum (like RHEL
and mandrake). They provide the "yum" capability, and that is what yume
requires. So distros which include yum will not install it. I had no
opportunity to build yum-oscar on fedora and mandrake, so I'm not sure they
build cleanly there. yum-oscar is depending on the python version, therefore
is distro (version) dependent, but not architecture dependent.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@3911 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 803a91bb1288c36502fc6c28c3afc1ffff76b08c