This script helps integrating images created outside the OSCAR GUI into the OSCAR/SIS infrastructure. For example: images created by yume. Usage: integrate_image --name testimg --arch ia64 --disktab /opt/oscar/oscarsamples/mydisktab
The image directory testimg must contain a valid image and be located inside
/var/lib/systemimager/images.

The default architecture attributed to the image is that of the master node. The
example explicitely selects ia64.

A disktable file should be inserted into the image by using the --disktab
command line option. This generates the /etc/systemimager/autoinstall.conf file
inside the image. This is not necessary if one has copied the image from another
OSCAR cluster.

A record is created in the SIS image database. No ODA entry is created, as that one
isn't used, yet. I think.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@3912 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: cdcdaecd87f84b580efb612f729899e09612c53a