Merged commit r4164 by efocht from 2006-01-25 14:04:02 +0100 (Wed, 25 Jan 2006) from repository https://svn.oscar.openclustergroup.org/svn/oscar/tmp/ef_sandbox to repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk Original commit message was:
Had to split this out of Package because it is needed during the prereqs
install which happens such early that we cannot be sure that XML::Simple
(required by OSCAR::Package) is already available. If all goes well, this
could replace the simple RPM_POOL variable provided by OSCAR::Package.

The package provides functions for building repository paths depending on
distro, version, etc... Currently only two functions are provided:

distro_repo_url($distro,$distro_version,$architecture)
------------------------------------------------------
The URL where distribution packages are stored. When the files are
stored locally they should go to:
/tftpboot/distro/$distro-$version-$arch
Doing this without subtrees of dirs has the advantage to immediately show
the admin how many of these disk space consuming distro repositories are
around.

If the nodes have connectivity to the internet one could use a publicly
accessible URL for the distro files. In that case place the URL of the
repository (yum or apt) into the file
/tftpboot/distro/$distro-$version-$arch.url


oscar_repo_path($distro,$distro_version,$architecture)
------------------------------------------------------
The path where OSCAR packages for a particular distro/version/arch
combination are stored. This path is defined as:
/tftpboot/oscar/$distro-$version-$arch

The distro and version names used are those detected as "compat_distro"
and "compat_distrover" in the OCA::OS_Detect framework. The reason is that
OSCAR doesn't care about the particular flavor of a rebuilt distro, it
uses the same packages eg. for rhel4, scientific linux 4 and centos 4.




git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4247 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: f15ec8a9dcb5e81cfac118428637bb9cebc20285