Merged commit r4172 by efocht from 2006-01-25 19:33:32 +0100 (Wed, 25 Jan 2006) from repository https://svn.oscar.openclustergroup.org/svn/oscar/tmp/ef_sandbox to repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk Original commit message was:
Added pattern matching to the prereq.cfg file parsing. And changed the
separator from "-" to ":" (redhat has "-" in its name :-( )

So a prereq.cfg file now looks like this:
[distro:version:architecture]
package_name_or_rpm_file
...
!package_name
...
[distro2:version:architecture]
...

The distro name, version or arch info can contain a "*". This matches
like a .* regexp.
Examples:
[redhat-el*:3:*] matches all rhel3 architectures.
[fedora:*:*]     matches all fedora core distros on all architectures
[mandr*:*:*]     matches both mandrake and mandriva
[*:*:*]          matches everything.

Attention! The real distro names are used here, not the compat names! This
makes it possible to distinguish between redhat-el-ws and centos.



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4253 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 64b9c166900826b97b2a910f33ead432fab81fd2