Merged commit r4218 by efocht from 2006-01-31 13:30:15 +0100 (Tue, 31 Jan 2006) from repository https://svn.oscar.openclustergroup.org/svn/oscar/tmp/ef_sandbox to repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk Original commit message was:
Trying to reduce the Package.pm functionality to run_pkg_scripts and
alike.
- moved some functions to Database.pm (functions dealing with listing opkgs)
- removed a bunch of obsolete stuff doing rpm copying, installing, deleting.
- removed routine for building rpmlists. This needs to use database functions
and was replaced by OSCAR::Database::pkgs_of_opkg which is much more flexible
in the selection of base distro/arch.
- removed old crap which required reading in config.xml and other XML files
which don't exist any more. This info goes to the database and there is no
need to maintain unnecessary duplicate code.
- outsourced some functionality to PackagePath.pm



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4295 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 7d8f3776c7340d12aa10e2f06f6e890de337e7bd