This is the rework of the build infrastructure I talked about at the phone conference.
Summary:
Instead of the sequence:
	./autogen.sh
	./configure
	make dist
you now just need to do
        make dist
The process is 10-20 times faster, depending on what you do.

automake and configure are gone. Also all Makefile.am are gone.

In order to ease testing and development one can now install directly the SVN
checkout by doing "make test".



If you want to test and develop directly on the OSCAR SVN do (inside the trunk
checkout):
   make test
This will set up the needed repositories, build the perl-Qt related files in
src but _not_ insert the copyright mesages and not build the docs. After this
you will be able to do:
   ./install_cluster eth0
from inside the trunk checkout.

If you finished your work you should try
   make clean
which will clean up the src directory and remove the checksum directory tmp/
such that your checkout will be as fresh as before.


If you want to build the OSCAR distribution you will need to do:
   make dist

This command will create a set of tarballs in the _parent_ directory of the
trunk checkout!

The OSCAR distibution comes in two OSCAR tarballs and a set of repository
tarballs:
  oscar-base-OSCARVERSION.tar.gz  : the base OSCAR tarball stripped of
                                    almost every binary packages (rpms)
  oscar-srpms-OSCARVERSION.tar.gz : the SRPMS, you only need this if you
                                    want to rebuild rpms
  oscar-repo-common-rpms-OSCARVERSION.tar.gz :
                                    repository of noarch rpms used on all
                                    supported distributions
  oscar-repo-DISTRO-VER-ARCH-OSCARVERSION.tar.gz :
                                    repository of distro and architecture
                                    specific rpms.

For an OSCAR installation:
 - Download the oscar-repo-common-rpms and the oscar-repo-DISTRO-VER-ARCH
 tarballls suitable for the distribution(s). Unpack them in /tftpboot/oscar.
 - Download the oscar-base tarball. Unpack it in an appropriate directory.
 Rename the unpacked directory, if you want (e.g. /opt/oscar)
 - In the unpacked directory (e.g. /opt/oscar) execute:
 ./install_cluster ETH_INTERFACE

 (ETH_INTERFACE is the cluster internal interface on the head node,
 e.g. eth0).




git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4831 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 64f2dc0cbf1b873741c72a8d72df8d21ff83966a