An attempt to fix some issues with start-over. Checking for services in the oscar rpms, though I believe this is not needed. When deinstalling rpms the services will be stopped anyway. Excluding the oda package from deinstallation, this keeps mysql around, as requested by DongInn.
This is certainly buggy and not working. Please fix the bugs yourselves. With
the new structure the old start_over stopped to work, so this is the best you
get right now. But with yume-opkg one can almost do the entire removal job a
lot cleaner... Removal of prereqs is missing!


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4858 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: f975312df17230cfeb3dca33461e1ed3ffe9708d