Update sync_files to a new version (which evolved in my repository).
New feature: it supports distro-templates. Look at /opt/sync_files/templates,
the directories and files therein. The templates should contain the system
accounts for that particular system. They will be filtered out from the
corresponding files on the master and appropriate templates will be added for
the targetted distribution. This allows a RHEL master to have SUSE client
nodes.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4924 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 8cb00343e350d89bb304be29d2cedfcd560196db