Updating yum-oscar to version 2.4.3. Unfortunately this required the build of sqlite,python-sqlite, python-urlgrabber on FC3 and original RHEL3/4. I don't have a RHEL3-i386 system, so those rpms are missing and someone will need to rebuild them.
Updated yume to version 2.1-1. It includes now yum-repoquery, there is no need
to install yum-utils any more. Removed yum-utils and dependency of it.

Moved python-elementtree from apitest to yume. This is because we should be
able to bootstrap a oscar-base system for getting it to run with yum(e) even
without having access to a repository.

Adapted the build.cfg files somewhat. Untested.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@4936 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: c23a1d086827dbf0b69d610e928294291fbbec6b