Moved per opkg configuration variables to the OSCAR database. This means: .configurator.values files are not created any more! The values go to the Packages_config table in the OSCAR database.
The Configbox.pm module was modified a bit to access the configurator
variables from the ODA. Also some packages' API scripts needed changes. I
tested this to work for ganglia and ntpconfig, I hope I cought all scripts
which needed to be changed. If you find anything which still needs to be
changed, please tell me.

Configbox.pm takes care of the configurator values handling. Use
   use OSCAR::Configbox;
   my $values = readInConfigValues($cfile,$opkg,$context,[noarray=>1]);

The noarray arg is optional.
You'll get back a hash reference, as you were used to.

If you want to interact directly with the database use the routines:
  set_pkgconfig_var
  get_pkgconfig_vars
  del_pkgconfig_vars
  pkgconfig_values
from OSCAR::Database.pm (this is addressed to Wesley for CLI).

Otherwise nothing changed.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@5066 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 827cded4d4a0e17f94c2ba534f509f00d79d45ce