Merged commit r5220 by efocht from 2006-07-24 19:35:52 +0200 (Mon, 24 Jul 2006) from repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk to repository https://svn.oscar.openclustergroup.org/svn/oscar/branches/branch-5-0 Original commit message was:
Distro-query returns now a preprocessed list of repositories if queried with
the "--node" option. Repositories local to the master node will get the string
"http://oscar_server/repo" prepended in the expectation that the repository is
exported. Which it better should be.
This is a problem with trunk and branch-5-0, therefore should be fixed in
branch-5-0, too.




git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-5-0@5230 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 6cffef5d7ca18cb42947f83a0544ad8e953b0404