Merged commit r5287 by efocht from 2006-08-04 18:12:38 +0200 (Fri, 04 Aug 2006) with svnmerge_1by1 version v1.1 31.7.2006 from repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk to repository https://svn.oscar.openclustergroup.org/svn/oscar/branches/branch-5-0 Original commit message was:
Fix for bug #244
The current (and latest) log file is oscarinstall.log. Old log files
get a number appended (eg. oscarinstall.log_27), starting with _1
and increasing with each new invocation of install_cluster.

Date stamp added to the beginning of the output such that the time of
invocation of install_cluster can be seen immediately.

Closes #244



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-5-0@5296 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 8f46cb9ba9cbea28ff5f16207618647862690e43