Merged commit r5371 by efocht from 2006-09-27 17:37:36 +0200 (Wed, 27 Sep 2006) with svnmerge_1by1 version v1.1 31.7.2006 from repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk to repository https://svn.oscar.openclustergroup.org/svn/oscar/branches/branch-5-0 Original commit message was:
Base package. This should be kept separate from the base prereq!
The base prereq needs to prepare the master node in an early phase
of the installation. It basically triggers the installation of
a bunch of necessary packages.
The base package is aimed at the client nodes. Currently it contains a
post_rpm_nochroot script which leads to the installation of $OSCAR_HOME/libs
and $OSCAR_HOME/scripts into each generated client image. It also copies
/etc/profile.d/oscar_home.{sh,csh} into the image. This is necessary in order
to get access to the database and OS_Detect functionality from the clients.

Part of bugfix for long lasting package In/Un issues (#33). Should be merged
into branch, too.



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-5-0@5391 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: c287acdf20dd6ca5460e132d4af36f797d7c1151