Merged commit r5531 by efocht from 2006-11-10 17:44:31 +0100 (Fri, 10 Nov 2006) with svnmerge_1by1 version v1.1 31.7.2006 from repository https://svn.oscar.openclustergroup.org/svn/oscar/trunk to repository https://svn.oscar.openclustergroup.org/svn/oscar/branches/branch-5-0 Original commit message was:
Improved version of post_rpm_nochroot. This one can be used to update the
systemconfig.conf files in an image after a kernel update. All you need to do
is:

cd $OSCAR_HOME/packages/sis/scripts
./post_rpm_nochroot --update /var/lib/systemimager/images/IMAGENAME

The tools scexec and scpush are used to push the files to the nodes
corresponding to the image.
Related to bug #284...



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/branch-5-0@5532 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: dc7bad68c08ba6b462f521a18561723711f31ce3