This opkg contains the RPMs for linux-ha. The heartbeat packages are installed on all nodes. The configuration is NOT done automatically, it depends a lot on what _exactly_ one wants to do with heartbeat.
Done as preparation for a GSoC project. It should focus on providing some
generic way of configuring reasonable heartbeat services.

NOTE:
Some packages are included because they are needed in order to build the
heartbeat package from SRPM (eg. swig, libnet).
heartbeat-ldirectord is not yet installed (in config.xml)! Once we have some
use for it, we should do it (makes sense only with Linux Virtual Server).


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@5746 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: b3db21e871cbefc20a6f144f0af7a91d83bdf09d