New schema for opkgc based OSCAR package management.
It makes little sense to carry the whole unnecessary garbage around.
Deleting the unneeded tables will immediately point us to the places
which need to be fixed.

The core opkg related tables are:
 Packages
 Node_Package_Status

Packages should actually be renamed to Opkgs.
It will contain the list of all opkgs visible in all accessible repositories.

The more I think about it, the less I believe Packages is needed at all.
We could populate it with the output of OpkgDB::opkg_list_available, but
those are only the packages available for the distribution of the master
node. That's not good!


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@6381 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 500143d8a19d3413e1e0999fabe3d04063af7006