Trimmed ODA schema. Removed tables:  Packages_requires  Packages_provides  Packages_conflicts This info is never needed. It is automatically resolved by the native package manager.
Trimmed Packages table. It is clearly overdesigned since the opkg-* can't
transport all information which this was designed for. Main change (will
need code revies): the fields name, summary, package, description can
not be _all_ filled with info. Left "name" and removed "package". This
is the opkg name. Summary and description can be filled with info from
the native opkg-*. Added a distro field. This one should hold the target
distribution for the package. This string is the compatible distro
string, i.e. the name which a local oscar repository directory would have.
Example: rhel-5-x86_64.

Now packages can show up several times, once for each distro for which they
are registered.

This is a less invasive change that first attempt and keps the door open
for two alternative approaches.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/big_cleanup_5.1@6396 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 2eb86cef8436a7a80e6595c8a37419b1fd007049