Add new GUI for OPD2. Note that since OPD2 does not download any OPKG anymore (usage of online repositories), the GUI is named OSCAR Repository Manager (ORM).
Also note that this is a C++/Qt4 code. The reason i did not use perlQt are:
- the latest version of perlQt only support Qt < 3.3 which is a very old version of Qt, note 
compliant anymore with Qt tools shipped in the major Linux distributions.
- C++/Qt4 is one of the latest technology for the creation of new GUIs.
- because the script 'opd2' can be used in non-interactive mode, it is not necessary to use 
perl modules; therefore the usage of C++ does imply any conflict with the existing OSCAR code.

At the same time, this is only a first implementation i quickly implemented (i did already 
know pretty weel C++/Qt); if need it is still possible to reuse .ui files and create the GUI 
in perlQt (or any other language that can be used with Qt).



git-svn-id: http://svn.oscar.openclustergroup.org/oscar/trunk@6441 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: b0b7db7dbb883a97ce96e73b3b7034568b4fdfd7