Added functionality to Selector to manage package sets. - re-enabled the package sets manager panel - fixed it to produce copies of package sets and rename them - relies on a new function in Database.pm which renames groups. - selecting and editing different package sets is possible now. - calls OSCAR::Opkg::write_pgroup_files when closed, such that the   package group files for systeminstaller are created for each package set.
TODO:
Clean up database and get rid of the notion of "selected" package set. The
selected package set is the set which gets installed on the master node.
I believe this should be a superset of all packages selected and assigned
to nodes. Need logic to build this superset.
Problem: the superset can be created out of all packagesets we created. But
it does not mean that all of these will be installed! This problem
exists with the OPM system, too: we only know which packages need to be
installed on the master if we have already built images and assigned them,
therefore know which packages are in use in the cluster.


git-svn-id: http://svn.oscar.openclustergroup.org/oscar/branches/big_cleanup_5.1@6473 18f9d06c-6ae2-0310-8438-938a554feff9

Former-commit-id: 7154fee88e91ea06b98ee2e79b34a492bc1b64f9